Added new step to the config & cleanup
