+++ sites/public/lib/ApplicationConductor.ts
@@ -10,10 +10,10 @@ import StepDefinition from "./StepDefinition"
 import AlternateContactStep from "./AlternateContactStep"
 import LiveAloneStep from "./LiveAloneStep"
 import HouseholdMemberStep from "./HouseholdMemberStep"
-import LiveWorkPreferenceStep from "./LiveWorkPreferenceStep"
-import DisplacedPreferenceStep from "./DisplacedPreferenceStep"
+// import LiveWorkPreferenceStep from "./LiveWorkPreferenceStep"
+// import DisplacedPreferenceStep from "./DisplacedPreferenceStep"
 import SelectedPreferencesStep from "./SelectedPreferencesStep"
-import PreferencesStep from "./PreferencesStep"
+import PreferencesStartStep from "./PreferencesStartStep"
 
 export const loadApplicationFromAutosave = () => {
   if (typeof window != "undefined") {
@@ -90,15 +90,7 @@ export default class ApplicationConductor {
     },
     preferencesStart: {
       url: "/applications/preferences/start",
-      definition: PreferencesStep,
-    },
-    preferencesLiveWork: {
-      url: "/applications/preferences/live-work",
-      definition: LiveWorkPreferenceStep,
-    },
-    preferencesDisplaced: {
-      url: "/applications/preferences/displaced",
-      definition: DisplacedPreferenceStep,
+      definition: PreferencesStartStep,
     },
     generalPool: {
       url: "/applications/preferences/general",
@@ -216,6 +208,7 @@ export default class ApplicationConductor {
   }
 
   routeToNextOrReturnUrl(url?: string) {
+    console.log("next url", this.determineNextUrl())
     void Router.push(lRoute(this.nextOrReturnUrl(url))).then(() => {
       this.returnToReview = false
       window.scrollTo(0, 0)
@@ -239,8 +232,11 @@ export default class ApplicationConductor {
   }
 
   determineNextUrl() {
+    console.log("determine next", this.currentStepIndex)
     let i = this.currentStepIndex + 1
     let nextUrl = ""
+
+    console.log("next url: ", this.steps[i])
 
     while (i < this.steps.length) {
       const nextStep = this.steps[i]
