+++ sites/public/lib/PreferencesStartStep.ts
@@ -1,0 +1,13 @@
+import StepDefinition from "./StepDefinition"
+
+export default class PreferencesStartStep extends StepDefinition {
+  skipStep() {
+    return !this.conductor.listing.preferences.length
+  }
+
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  save(formData: Record<string, any>) {
+    this.application.preferences = formData
+    this.conductor.sync()
+  }
+}
