--- sites/public/lib/PreferencesStep.ts
@@ -1,7 +1,0 @@
-import StepDefinition from "./StepDefinition"
-
-export default class PreferencesStep extends StepDefinition {
-  skipStep() {
-    return !this.conductor.listing.preferences.length
-  }
-}
